move makeIsotropicElasticityTensor in its dedicated file
